move things around for better exports
