+++ packages/ts-etl/package.json
@@ -9,7 +9,7 @@
     "clean": "rm -rf dist/",
     "test:integration": "ava test/integration/",
     "test:unit": "ava test/unit/",
-    "start": "node --loader=ts-node/esm --experimental-specifier-resolution=node --nolazy -r ts-node/register/transpile-only src/index.ts",
+    "start": "node --loader=ts-node/esm --experimental-specifier-resolution=node --nolazy -r ts-node/register/transpile-only src/cli.ts",
     "lint": "npx eslint .",
     "lint:fix": "npx eslint --fix .",
     "prepare": "tsc --declaration",
